Use Double for fabric roll quantity and refine quantity input

+++ core/base/src/commonMain/kotlin/com/erp/jytextile/core/base/extension/NumberExtensions.kt
@@ -2,6 +2,6 @@ package com.erp.jytextile.core.base.extension
 
 private const val YARD_PER_METER = 1.09361
 
-fun meterToYard(meter: Int): Int {
-    return (meter * YARD_PER_METER).toInt()
+fun meterToYard(meter: Double): Double {
+    return (meter * YARD_PER_METER)
 }
